+++ src/components/pages/index/codeViewer/CodeViewer.stories.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from 'react';
-import { constructCss } from '../../../fixtures/functions/constructSourceCode';
-import { getFlexboxPropertyInfoById } from '../../../fixtures/functions/extractData';
-import { useClipboard } from '../../../fixtures/hooks/useClipboard';
-import { Language } from '../../providers/CustomIntlProvider';
+import { constructCss } from '../../../../fixtures/functions/constructSourceCode';
+import { getFlexboxPropertyInfoById } from '../../../../fixtures/functions/extractData';
+import { useClipboard } from '../../../../fixtures/hooks/useClipboard';
+import { Language } from '../../../providers/CustomIntlProvider';
 import { StyledComponent as CodeViewer } from './CodeViewer';
 import { FileExtension, FILE_EXTENSION, getSourceCode } from './useCodeViewerState';
 export default {
