+++ src/components/pages/index/codeViewer/CodeViewerContent.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Light as SyntaxHighlighter } from 'react-syntax-highlighter';
 import monoBlue from 'react-syntax-highlighter/dist/esm/styles/hljs/mono-blue';
 import styled from 'styled-components';
-import { deviceMaxWidth } from '../../../data/deviceSize';
+import { deviceMaxWidth } from '../../../../data/deviceSize';
 import { FileExtension } from './useCodeViewerState';
 
 type Props = {
