+++ src/components/pages/index/codeViewer/CodeViewerHeader.tsx
@@ -1,11 +1,11 @@
 import { faBook, faCopy } from '@fortawesome/free-solid-svg-icons';
 import React from 'react';
 import styled from 'styled-components';
-import { createReferenceUrl } from '../../../fixtures/functions/reference';
-import { FontAwesomeIconButton } from '../../common/button/FontAwesomeIconButton';
-import { TextButton, TEXT_BUTTON_THEME } from '../../common/button/TextButton';
-import { FontAwesomeIconLink } from '../../common/link/FontAwesomeIconLink';
-import { Language } from '../../providers/CustomIntlProvider';
+import { createReferenceUrl } from '../../../../fixtures/functions/reference';
+import { FontAwesomeIconButton } from '../../../common/button/FontAwesomeIconButton';
+import { TextButton, TEXT_BUTTON_THEME } from '../../../common/button/TextButton';
+import { FontAwesomeIconLink } from '../../../common/link/FontAwesomeIconLink';
+import { Language } from '../../../providers/CustomIntlProvider';
 import { FileExtension, FILE_EXTENSION } from './useCodeViewerState';
 
 type Props = {
