+++ src/components/pages/index/codeViewer/useCodeViewerState.ts
@@ -1,9 +1,9 @@
 import { useContext, useState } from 'react';
-import { constructCss, constructHtml } from '../../../fixtures/functions/constructSourceCode';
-import { getFlexboxPropertyInfoById } from '../../../fixtures/functions/extractData';
-import { CustomIntlContext, Language } from '../../providers/CustomIntlProvider';
-import { OpenedCodeViewerContext } from '../../providers/OpenedCodeViewer';
-import { SelectedFlexboxPropertyIdContext } from '../../providers/SelectedFlexboxPropertyIdProvider';
+import { constructCss, constructHtml } from '../../../../fixtures/functions/constructSourceCode';
+import { getFlexboxPropertyInfoById } from '../../../../fixtures/functions/extractData';
+import { CustomIntlContext, Language } from '../../../providers/CustomIntlProvider';
+import { OpenedCodeViewerContext } from '../../../providers/OpenedCodeViewer';
+import { SelectedFlexboxPropertyIdContext } from '../../../providers/SelectedFlexboxPropertyIdProvider';
 
 export type FileExtension = typeof FILE_EXTENSION[keyof typeof FILE_EXTENSION];
