+++ src/components/pages/index/flexbox/FlexboxGroup.stories.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { getFlexboxProperty } from '../../../fixtures/functions/extractData';
+import { getFlexboxProperty } from '../../../../fixtures/functions/extractData';
 import { FlexboxGroup } from './FlexboxGroup';
 
 export default {
